web/requirements.txt
akismet==0.2.0
appdirs==1.4.0
beautifulsoup4==4.5.3
click==6.7
Django==1.10.5
django-appconf==1.0.2
django-debug-toolbar==1.6
django-disqus==0.5
django-extensions==1.7.6
django-grappelli==2.9.1
django-haystack==2.6.0
django-imagekit==3.3
django-pagination==1.0.7
django-profiles2==0.7
django-registration-redux==1.4
django-reversion==2.0.8
django-rosetta==0.7.12
django-taggit==0.22.0
django-taggit-templatetags2==1.6.1
django-tastypie==0.13.3
django-templatetag-sugar==1.0
django-webtest==1.8.0
feedparser==5.2.1
first==2.0.1
funcsigs==1.0.2
html5lib==0.90
lxml==3.7.2
microsofttranslator==0.8
mock==2.0.0
olefile==0.44
packaging==16.8
pbr==1.10.0
pilkit==1.1.13
Pillow==4.0.0
polib==1.0.8
psycopg2==2.6.2
pyparsing==2.1.10
python-dateutil==2.6.0
python-mimeparse==1.6.0
reportlab==3.3.0
requests==2.13.0
six==1.10.0
sqlparse==0.2.2
waitress==1.0.2
webencodings==0.5
WebOb==1.7.1
WebTest==2.0.25
Whoosh==2.7.4
git+https://github.com/pando85/django-navbar
git+https://github.com/detrout/django-relationships
git+https://github.com/wiledion/django-ratings

